requirements.txt
ultralytics
matplotlib
pyyaml
tqdm